fix: divided the header into components
