+++ src/components/common/Header/NavBar/NavBar.module.scss
@@ -1,0 +1,43 @@
+@import '@/theme/variables';
+@import '@/theme/mixins';
+
+.nav {
+  display: flex;
+  @include text-small;
+
+  .navList {
+    list-style-type: none;
+    text-decoration: none;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+  .navLink {
+    color: $white-text;
+    text-decoration: none;
+    padding-inline: 1.5rem;
+    position: relative;
+    padding-bottom: 3px;
+
+    &::after {
+      content: '';
+      position: absolute;
+      left: 20px;
+      bottom: 0;
+      width: 0;
+      height: 1px;
+      background-color: $background-two;
+      transition: width 0.3s cubic-bezier(0.45, 0.05, 0.55, 0.95);
+    }
+
+    &:hover::after {
+      width: 68%;
+    }
+  }
+
+  @media (max-width: $tablet) {
+    & {
+      display: none;
+    }
+  }
+}
